+++ crabml-core/src/tokenizer/mod.rs
@@ -1,4 +1,5 @@
 mod tokenizer;
+mod tokenizer_gpt2;
 mod tokenizer_llama;
 
 pub use tokenizer::BpeTokenizer;
