+++ crabml-core/src/tokenizer/tokenizer.rs
@@ -1,5 +1,4 @@
 use std::cell::RefCell;
-use std::collections::HashMap;
 use std::rc::Rc;
 
 use super::tokenizer_llama::LlamaTokenEncoder;
